chore(OrderSuccess): Use new Member.publicMessage
Also moved part of the logic to subcomponent to make the code clearer and avoid unnecessary renders.
